requirements.txt
arrow==1.2.2
blis==0.7.7
boto3==1.23.8
botocore==1.26.8
buddy==0.0.3
catalogue==1.0.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cloudpickle==2.1.0
cymem==2.0.6
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
gutils==0.1.9
gym==0.23.1
gym-notices==0.0.6
h5py==3.7.0
hashids==1.3.1
idna==3.3
importlib-metadata==4.11.4
jericho==3.1.0
jmespath==1.0.0
mementos==1.3.1
more-itertools==8.13.0
murmurhash==1.0.7
networkx==2.8.2
numpy==1.22.4
plac==1.1.3
preshed==3.0.6
prompt-toolkit==3.0.29
pycparser==2.21
python-dateutil==2.8.2
PyYAML==6.0
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
spacy==2.3.7
srsly==1.0.5
tabulate==0.8.9
TatSu==5.6.1
textworld==1.5.2
thinc==7.4.5
torch==1.11.0
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
wasabi==0.9.1
wcwidth==0.2.5
zipp==3.8.0
